AEGIS-6406 - moved scripts
