+++ requirements.txt
@@ -1,3 +1,2 @@
 torch>=1.7,<2
-click>=7.1,<8
 r2pipe>=1.5,<2
